Refactor weather data handling and integrate OpenMeteo API
Replaced WeatherTemplateDbContext with WeatherDbContext and refactored weather-related database logic. Introduced OpenMeteoService for external API calls, enabling enriched weather data retrieval. Updated project structure and dependencies to support the new service architecture.

+++ DTO/Output/GetWeatherByLocationOutputDto.cs
@@ -2,7 +2,7 @@
 
 public class GetWeatherByLocationOutputDto
 {
-    public int Temperature { get; init; }
+    public double Temperature { get; init; }
     public int WindDirection { get; init; }
     public int WindSpeed { get; init; }
     public DateTime SunriseDateTimeIso8601 { get; init; }
